correction mathilde + arrondi benef et parité + logs pour amende et boost productivité

+++ agt/msgTypes.go
@@ -3,10 +3,10 @@ package agt
 type SituationActuelle struct {
 	NbEmp  int     `json:"nbEmp"`
 	Parite float64 `json:"parite"`
-	Benef  float64 `json:"benefices"`
+	Benef  int     `json:"benefices"`
 }
 
-func NewSituationActuelle(nbemp int, parite float64, benef float64) *SituationActuelle {
+func NewSituationActuelle(nbemp int, parite float64, benef int) *SituationActuelle {
 	return &SituationActuelle{
 		NbEmp:  nbemp,
 		Parite: parite,
